+++ src/OutputFormat.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the RegexParser package.
+ *
+ * (c) Younes ENNAJI <younes.ennaji.pro@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace RegexParser;
+
+/**
+ * Output formats accepted by Regex::explain() and Regex::highlight().
+ */
+enum OutputFormat: string
+{
+    case TEXT = 'text';
+    case HTML = 'html';
+    case CONSOLE = 'console';
+}